third_party/blink/web_tests/virtual/omt-worker-fetch/external/wpt/workers/baseurl/alpha/xhr-in-sharedworker-expected.txt

This is a testharness.js-based test.
FAIL xhr-worker assert_equals: expected "gamma\n" but got "beta\n"
Harness: the test ran to completion.
